+++ sn_faucet/src/faucet_server.rs
@@ -36,11 +36,13 @@ pub async fn run_faucet_server(client: &Client) -> Result<()> {
     let server =
         Server::http("0.0.0.0:8000").map_err(|err| eyre!("Failed to start server: {err}"))?;
     claim_genesis(client).await.map_err(|err| {
+        println!("Faucet Server couldn't start as we failed to claim Genesis");
         eprintln!("Faucet Server couldn't start as we failed to claim Genesis");
         error!("Faucet Server couldn't start as we failed to claim Genesis");
         err
     })?;
 
+    // This println is used in sn_testnet to wait for the faucet to start.
     println!("Starting http server listening on port 8000...");
     debug!("Starting http server listening on port 8000...");
     for request in server.incoming_requests() {