+++ sn_cli/CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.86.36](https://github.com/maidsafe/safe_network/compare/sn_cli-v0.86.35...sn_cli-v0.86.36) - 2023-12-05
+
+### Other
+- *(networking)* remove triggered bootstrap slowdown
+
+## [0.86.35](https://github.com/maidsafe/safe_network/compare/sn_cli-v0.86.34...sn_cli-v0.86.35) - 2023-12-04
+
+### Other
+- update dependencies
+
+## [0.86.34](https://github.com/maidsafe/safe_network/compare/sn_cli-v0.86.33...sn_cli-v0.86.34) - 2023-12-01
+
+### Other
+- *(ci)* fix CI build cache parsing error
+
 ## [0.86.33](https://github.com/maidsafe/safe_network/compare/sn_cli-v0.86.32...sn_cli-v0.86.33) - 2023-11-29
 
 ### Other