+++ sn_networking/CHANGELOG.md
@@ -6,6 +6,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.1](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.11.0...sn_networking-v0.11.1) - 2023-12-05
+
+### Other
+- *(node)* refactor NetworkEvent handling
+- *(network)* allow replication even below K_VALUE peers
+- *(networking)* dont resort closest peers list
+- tie node reward test to number of data.
+- *(networking)* remove triggered bootstrap slowdown
+- *(networking)* remove extended spend wait before verification
+- log swarm.NetworkInfo
+- log on query
+
+## [0.11.0](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.10.27...sn_networking-v0.11.0) - 2023-12-01
+
+### Added
+- *(network)* use seperate PUT/GET configs
+
+### Other
+- *(ci)* fix CI build cache parsing error
+- *(network)* [**breaking**] use the Quorum struct provided by libp2p
+
 ## [0.10.27](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.10.26...sn_networking-v0.10.27) - 2023-11-29
 
 ### Added