+++ sn_node/CHANGELOG.md
@@ -2365,6 +2365,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - *(node)* increase interval for interval_replication
 - *(networking)* reduce nodes receiving replication updates.
 
+## [0.99.0](https://github.com/maidsafe/safe_network/compare/sn_node-v0.98.44...sn_node-v0.99.0) - 2023-12-01
+
+### Added
+- *(network)* use seperate PUT/GET configs
+
+### Other
+- *(ci)* fix CI build cache parsing error
+- *(network)* [**breaking**] use the Quorum struct provided by libp2p
+
+## [0.99.1](https://github.com/maidsafe/safe_network/compare/sn_node-v0.99.0...sn_node-v0.99.1) - 2023-12-04
+
+### Other
+- update dependencies
+
+## [0.99.2](https://github.com/maidsafe/safe_network/compare/sn_node-v0.99.1...sn_node-v0.99.2) - 2023-12-05
+
+### Other
+- *(node)* update assert over royalty count
+- *(node)* refactor NetworkEvent handling
+- *(network)* allow replication even below K_VALUE peers
+- tie node reward test to number of data.
+- FORWARDER_CHOOSING_FACTOR docs
+- tie verification data test to repl time
+- *(node)* separation of cheap NetworkEvent handling
+
 ## v0.1.0 (2023-05-04)
 
 ### Chore