+++ sn_testnet/CHANGELOG.md
@@ -2100,6 +2100,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other
 - update dependencies
 
+## [0.2.316](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.315...sn_testnet-v0.2.316) - 2023-12-01
+
+### Other
+- update dependencies
+
+## [0.2.317](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.316...sn_testnet-v0.2.317) - 2023-12-04
+
+### Added
+- *(testnet)* wait till faucet server starts
+
+### Fixed
+- *(testnet)* dont be taking faucet stdout
+
+## [0.2.318](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.317...sn_testnet-v0.2.318) - 2023-12-05
+
+### Other
+- *(networking)* remove triggered bootstrap slowdown
+
 ## v0.1.0 (2023-03-16)
 
 <csr-id-4f04bd1a5d1c747bfc6b5d39824dd108f8546b7b/>